Add files via upload
In this project, we are directed to develop a graphical user interface (GUI) for a system that stores data in the array list and to add a developed GUI with the project we completed in the first coursework where we developed a parent class i.e, Vehicle class and two child classes (AutoRickshaw and ElectricScooter class). The main objective of this coursework is to make us familiar with developing Graphical User Interface on java and to store the detail and data on an array list. This coursework consists of a total of four classes, three classes from previous coursework and one which we developed during this coursework. I developed a TransportGUI class for the graphical user interface. Some methods were created including the main method in this class. Various dialogue boxes are also created to display information and data through GUI.